refactor: refactoring custom component
