--- components/border.tsx
@@ -1,5 +1,0 @@
-const Border = () => {
-  return <div className="border-gray"></div>;
-};
-
-export default Border;
