+++ components/custom/check-box.tsx
@@ -1,0 +1,7 @@
+interface Props {}
+
+const CheckBox = (props: Props) => {
+  return <div></div>;
+};
+
+export default CheckBox;
