+++ components/custom/index.ts
@@ -1,0 +1,4 @@
+export { default as CustomInput } from "./input";
+export { default as CustomCheckBox } from "./check-box";
+export { default as CustomNicknameEditForm } from "./nickname-edit-form";
+export { default as CustomSelectBox } from "./select-box";
