--- components/onBoarding.tsx
@@ -1,3 +1,0 @@
-export default function OnBoardingPage() {
-  return <h1>onBoardingPage</h1>;
-}
