+++ styles/globals.css
@@ -8,6 +8,7 @@ body {
   height: 100%;
   font-family: Pretendard, -apple-system, BlinkMacSystemFont, Segoe UI, Roboto,
     Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+  background: var(--bg);
 }
 
 a {
@@ -19,6 +20,7 @@ a {
   box-sizing: border-box;
 
   --primary: #ECA354;
+  --bg: #FBFBFB;
 }
 
 .border-gray {
