+++ styles/layout.module.css
@@ -1,11 +1,14 @@
 .wrapper {
-  max-width: 480px;
   width: 100%;
   height: 100%;
-  position: relative;
+  max-width: 480px;
+  overflow: auto;
+  padding: 56px 0;
 }
 
 .header {
+  max-width: 480px;
+  top: 0;
   width: 100%;
   height: 56px;
   box-shadow: 0px 2px 4px rgba(203, 203, 203, 0.25);
@@ -15,8 +18,8 @@
   font-weight: 700;
   transition: 0.6s ease-in-out;
   position: fixed;
- z-index: 1000; 
- background: #fff;
+  z-index: 1000; 
+  background: #fff;
 }
 
 .header:active {
@@ -26,6 +29,7 @@
 .nav {
   background-color: white;
   position: fixed;
+  max-width: 480px;
   bottom: 0;
   width: 100%;
   box-shadow: 0px -2px 4px rgba(203, 203, 203, 0.25);
@@ -66,6 +70,5 @@
 .main{
   width: 100%;
   height: 100%;
-  overflow: auto;
-  padding: 56px 0;
+  padding: 24px 16px;
 }
